Split project settings between original and public settings pages add public settings page
